requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow

celery==5.2.7
redis==4.3.3
django-celery-beat
django-celery-results

django-filter
python-decouple
djangorestframework_simplejwt
python-slugify
wheel
drf-yasg

gunicorn
django-cors-headers
whitenoise

aiogram

# django-rest-framework-social-oauth2

drf_social_oauth2
